+++ package.json
@@ -12,18 +12,18 @@
     "@hookform/resolvers": "^4.1.3",
     "@neondatabase/serverless": "^0.10.4",
     "@svgr/webpack": "^8.1.0",
-    "@tanstack/react-query": "^5.79.0",
+    "@tanstack/react-query": "^5.79.2",
     "axios": "^1.9.0",
     "next": "15.0.4",
-    "next-auth": "^4.24.11",
+    "next-auth": "5.0.0-beta.28",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-hook-form": "^7.56.4",
-    "zod": "^3.25.42",
+    "react-hook-form": "^7.57.0",
+    "zod": "^3.25.49",
     "zustand": "^5.0.5"
   },
   "devDependencies": {
-    "@tanstack/react-query-devtools": "^5.79.0",
+    "@tanstack/react-query-devtools": "^5.79.2",
     "@trivago/prettier-plugin-sort-imports": "^5.2.2",
     "@types/eslint": "^8.56.12",
     "@types/node": "^20.17.57",
@@ -37,5 +37,5 @@
     "tailwindcss": "^3.4.17",
     "typescript": "^5.8.3"
   },
-  "packageManager": "pnpm@10.11.0"
+  "packageManager": "pnpm@10.11.1"
 }
