+++ src/providers/SessionProvider.tsx
@@ -9,9 +9,15 @@ interface AuthSessionProps {
   children: ReactNode;
 }
 
-//TODO - 수정
 const AuthSession = ({ session, children }: AuthSessionProps) => {
-  return <SessionProvider session={session}>{children}</SessionProvider>;
+  return (
+    <SessionProvider
+      session={session}
+      basePath="/api/auth"
+    >
+      {children}
+    </SessionProvider>
+  );
 };
 
 export default AuthSession;
