+++ src/types/next-auth.d.ts
@@ -3,10 +3,12 @@ import "next-auth/jwt";
 
 declare module "next-auth" {
   interface Session {
-    accessToken: string;
+    accessToken?: string;
     expires: string;
-    socialId: number;
     status: string;
+  }
+  interface User {
+    socialId: string;
   }
 }
